requirements.txt
pygame~=2.0.1
numpy~=1.20.1
pyserial~=3.5